incremental cache updating prototype
this is completely ignorant of re-issued data, so when a new issue of data happens, the cache may be somewhat stale and/or inaccurate.  the inaccuracies will probably be relatively small if the reissues are not wildly different than the previous issue.  also, those differences will get washed out as data series continue to grow in size.  a ~weekly full population of the cache should probably be done to clean up those inaccuracies.

this will break ingestion if new data is loaded before fully populating the cache with the existing data, because it expects a new field in the cache: 'num_points'.

the code is currently saving results to a 'test' cache table for comparing with the regular/main cache table.  TODOs have been included for lines that need to be reverted if/when testing is successful.

i have a feeling that the incremental changes could result in some loss of precision for the sd values, but i may just be paranoid.  theoretically, the results should be identical if floats have infinite precision.

changed 'Database.get_covidcast_meta()' to 'Database.compute_covidcast_meta()' for clarity