allow state-level FIPS codes for geo_type "county"
this change is intended to remedy the empty state visualization on the covidcast dashboard (ie, as seen on https://delphi.cmu.edu/covidcast/?date=20221115&region=PA) that was noticed today.  

it should make it so state-level FIPS codes are also allowed as valid values for `geo_type` of "`county'" (ie, all of Pennsylvania is captured in FIPS code 42000 where individual counties can be found in the range of 42001-42999).  it looks like "`hhs`" would work identically in place of "`chng-fips`" too, which kind of implies "hhs" and "county" may be redundant geo_types.

this is also largely untested