Creates Youtube-survey doc page
This is a draft of the Youtube-survey doc page. A lot of information is missing or assumed based off of (https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/fb-survey.html) and may very possibly altogether be incorrect. A comment from Katie from a May 4/7 2020 tooling team sprint states: "YouTube 4-question survey - this is essentially a copy of fb-survey without the weighted signals".